mystery-imagegen/requirements.txt
Flask==2.0.1
Werkzeug<2.1.0
requests==2.28.1
python-dotenv==0.19.2
Pillow>=10.0.0
markdown==3.4.1
markdown-it-py==2.2.0
